Grapher: Add Grapher class and improve main function
A new Grapher class has been added to handle the graphing of input functions in braille. This is accompanied by the inclusion of Grapher headers in relevant files. Main function in main.cpp has been thoroughly modified, including the addition of an inputParser function and removal of some unnecessary code line.
 For now, main function is only a proof of concept.